Combine an offline LLM with voice recognition and text-to-speech.
Using OLlama with qwen3:4b and a custom voice, made an offline AI-Assistant who will answer a question that is said through the microphone. In the tests I used Spanish and a time of 10 seconds to say a message. This commit won't include the model for the tts as it's heavy.

Also I used asyncio module to improve time response in both the hearing and reading process.